// sd-photobooth_GIT/types/type.ts
export type Location = {
  id: number;
  locationName: string;
  address: string;
  createdAt: string;
  lastModified: string;
  createdById: string | null;
  lastModifiedById: string | null;
};